NL-to-Graphql/Schema-rag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.212.0
chromadb==0.4.18
openai==1.109.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic==2.12.5

# LangChain
langchain==1.2.3
langchain-core==1.2.7
langchain-community==0.4.1
langchain-openai==1.1.7
langchain-groq==1.1.1
langchain-text-splitters==1.1.0
langchain-classic==1.0.1
langsmith==0.6.2

# LangGraph
langgraph==1.0.5
